+++ packages/matter-node-shell.js/src/MatterNode.ts
@@ -23,7 +23,10 @@ export class MatterNode {
     commissioningController?: CommissioningController;
     private matterDevice?: MatterServer;
 
-    constructor(private nodeNum: number) {}
+    constructor(
+        private readonly nodeNum: number,
+        private readonly netInterface?: string,
+    ) {}
 
     async initialize(resetStorage: boolean) {
         /**
@@ -85,7 +88,7 @@ export class MatterNode {
          * are called.
          */
 
-        this.matterDevice = new MatterServer(this.storageManager);
+        this.matterDevice = new MatterServer(this.storageManager, { mdnsInterface: this.netInterface });
         this.commissioningController = new CommissioningController({
             autoConnect: false,
         });