+++ packages/matter-node.js-examples/src/examples/DeviceNode.ts
@@ -124,7 +124,7 @@ class Device {
         const productName = `node-matter OnOff ${isSocket ? "Socket" : "Light"}`;
         const productId = getIntParameter("productid") ?? deviceStorage.get("productid", 0x8000);
 
-        const netAnnounceInterface = getParameter("announceinterface");
+        const netInterface = getParameter("netinterface");
         const port = getIntParameter("port") ?? 5540;
 
         const uniqueId = getIntParameter("uniqueid") ?? deviceStorage.get("uniqueid", Time.nowMs());
@@ -168,7 +168,7 @@ class Device {
          * are called.
          */
 
-        this.matterServer = new MatterServer(storageManager, { mdnsAnnounceInterface: netAnnounceInterface });
+        this.matterServer = new MatterServer(storageManager, { mdnsInterface: netInterface });
 
         const commissioningServer = new CommissioningServer({
             port,